dt changed to int from float

+++ UI2DObject.cpp
@@ -16,7 +16,7 @@ namespace FG
 
 	}
 
-	void UI2DObject::Update(float dt)
+	void UI2DObject::Update(int dt)
 	{
 
 	}
